Update: Styling updates to SDK site removed nav bar, updated font to NeueHaasUnica to match Waymark defaults, added headers to all pages
