+++ example_site/src/components/AccountAuthentication.js
@@ -153,7 +153,7 @@ function CreateAccountForm() {
 
 export default function AccountAuthentication() {
   return (
-    <>
+    <div className = 'center'>
       <Header 
         title="Create an Account or Log In"
         subtitle="To keep track of videos that users have created,
@@ -164,6 +164,6 @@ export default function AccountAuthentication() {
           <CreateAccountForm />
           <LoginAccountForm />
       </div>
-    </>
+    </div>
   );
 }
