+++ example_site/src/components/ConfigurationControls.css
@@ -24,7 +24,6 @@ button.configuration-submit-button {
   display: block;
   margin-top: 40px;
   text-align: center;
-  width: 12em;
 }
 
 .configuration-controls-subsection > label,
